+++ config_sample/hub.example.com.js
@@ -1,0 +1,5 @@
+vtcprefix = "example";
+company_distance_unit = "metric";
+vtccolor = "#888888";
+apidomain = "https://drivershub.charlws.com";
+navio_company_id = 0;