+++ config_sample/hub.example.com.php
@@ -1,0 +1,15 @@
+<?php
+    $vtcname = "Example";
+    $vtcabbr = "exp";
+    $vtccolor = "#888888";
+    $vtccolordark = "#666666";
+    $slogan = "This is a config sample";
+
+    $domain = "https://hub.example.com";
+    $domainpure = "hub.example.com";
+
+    $api = "drivershub.charlws.com";
+    $status = "status.charlws.com";
+
+    $enabled_plugins = ["announcement", "application", "division", "downloads", "event", "livemap", "ranking", "x_of_the_month"];
+?>